+++ drivers/src/nrfx_comp.c
@@ -28,7 +28,7 @@ static void comp_execute_handler(nrf_comp_event_t event, uint32_t event_mask)
     if (nrf_comp_event_check(event) && nrf_comp_int_enable_check(event_mask))
     {
         nrf_comp_event_clear(event);
-        NRFX_LOG_DEBUG("Event: %s.", (uint32_t)EVT_TO_STR(event));
+        NRFX_LOG_DEBUG("Event: %s.", EVT_TO_STR(event));
 
         m_comp_event_handler(event);
     }
@@ -52,8 +52,8 @@ ret_code_t nrfx_comp_init(nrfx_comp_config_t const * p_config,
     { // COMP driver is already initialized
         err_code = NRFX_ERROR_INVALID_STATE;
         NRFX_LOG_WARNING("Function: %s, error code: %s.",
-                         (uint32_t)__func__,
-                         (uint32_t)NRFX_LOG_ERROR_STRING_GET(err_code));
+                         __func__,
+                         NRFX_LOG_ERROR_STRING_GET(err_code));
         return err_code;
     }
 
@@ -70,8 +70,8 @@ ret_code_t nrfx_comp_init(nrfx_comp_config_t const * p_config,
     {
         err_code = NRFX_ERROR_INVALID_PARAM;
         NRFX_LOG_WARNING("Function: %s, error code: %s.",
-                         (uint32_t)__func__,
-                         (uint32_t)NRFX_LOG_ERROR_STRING_GET(err_code));
+                         __func__,
+                         NRFX_LOG_ERROR_STRING_GET(err_code));
         return err_code;
     }
 
@@ -80,8 +80,8 @@ ret_code_t nrfx_comp_init(nrfx_comp_config_t const * p_config,
     {
         err_code = NRFX_ERROR_BUSY;
         NRFX_LOG_WARNING("Function: %s, error code: %s.",
-                         (uint32_t)__func__,
-                         (uint32_t)NRFX_LOG_ERROR_STRING_GET(err_code));
+                         __func__,
+                         NRFX_LOG_ERROR_STRING_GET(err_code));
         return err_code;
     }
 #endif
@@ -124,9 +124,7 @@ ret_code_t nrfx_comp_init(nrfx_comp_config_t const * p_config,
     m_state = NRFX_DRV_STATE_INITIALIZED;
 
     err_code = NRFX_SUCCESS;
-    NRFX_LOG_INFO("Function: %s, error code: %s.",
-                  (uint32_t)__func__,
-                  (uint32_t)NRFX_LOG_ERROR_STRING_GET(err_code));
+    NRFX_LOG_INFO("Function: %s, error code: %s.", __func__, NRFX_LOG_ERROR_STRING_GET(err_code));
     return err_code;
 }
